separate releases vs inbetweens wrt package repo and docker tag
the main problem we solve here is that currently, people just install
the latest package, or run the latest docker image, and occasionally
something in master hasn't had the chance yet to be battle tested in
production.  We want to make it safer for them to run metrictank,
by guiding them towards the latest stable releases, which are tagged
after we feel more confident about the code in master.

while we're at it, make sure we consistently set version/tag everywhere